+++ merlin/models/torch/schema.py
@@ -1,0 +1,385 @@
+import types
+from typing import Callable, Dict, Optional, Tuple, TypeVar, Union
+
+import torch
+from torch import nn
+
+from merlin.dispatch.lazy import LazyDispatcher
+from merlin.schema import ColumnSchema, Schema, Tags
+
+Selection = Union[Schema, ColumnSchema, Callable[[Schema], Schema], Tags]
+ToSelectT = TypeVar("ToSelectT")
+
+
+class _LazyDispatchPyTorch(LazyDispatcher):
+    def __init__(self, func_or_name):
+        super().__init__(func_or_name)
+        self._tensor_registry = {}
+
+    def register_tensor(self, tensor_type, func=None):
+        if func is None:
+            return lambda f: self.register_tensor(tensor_type, f)
+        self._tensor_registry[tensor_type] = func
+
+        return func
+
+    def tensors(self, inputs):
+        for tensor_type, func in self._tensor_registry.items():
+            if torch.jit.isinstance(inputs, tensor_type):
+                return func(inputs)
+
+        raise NotImplementedError(f"Schema not registered for {type(inputs)}")
+
+    def get_schema(self, inputs: Union[torch.Tensor, Dict[str, torch.Tensor], Schema]) -> Schema:
+        if isinstance(inputs, Schema):
+            return inputs
+        return self.tensors(inputs)
+
+
+class _InputSchemaDispatch(_LazyDispatchPyTorch):
+    def __call__(self, module: nn.Module, inputs: Optional[Schema] = None) -> Schema:
+        if hasattr(module, "input_schema"):
+            output = module.input_schema
+            if isinstance(output, types.MethodType):
+                return output()
+
+            return output
+
+        input_schemas = getattr(module, "__input_schemas", None)
+        if input_schemas:
+            if len(input_schemas) == 1:
+                return input_schemas[0]
+
+            if inputs is None:
+                raise ValueError("Must provide inputs to get output schema")
+
+            # TODO: Fix this properly
+            i = module.__input_schemas.index(inputs)
+
+            return input_schemas[i]
+
+        try:
+            return super().__call__(module, inputs)
+        except NotImplementedError:
+            raise ValueError(
+                f"Could not get output schema of {module} " "please call mm.trace_schema first."
+            )
+
+    def trace(
+        self, module: nn.Module, inputs: Union[torch.Tensor, Dict[str, torch.Tensor], Schema]
+    ) -> Schema:
+        inputs_schema = self.get_schema(inputs)
+
+        try:
+            return super().__call__(module, inputs_schema)
+        except NotImplementedError:
+            return inputs_schema
+
+
+class _OutputSchemaDispatch(_LazyDispatchPyTorch):
+    def __call__(self, module: nn.Module, inputs: Optional[Schema] = None) -> Schema:
+        if hasattr(module, "output_schema"):
+            output = module.output_schema
+            if isinstance(output, types.MethodType):
+                return output()
+
+            return output
+
+        output_schemas = getattr(module, "__output_schemas", None)
+        if output_schemas:
+            if len(output_schemas) == 1:
+                return output_schemas[0]
+
+            if inputs is None:
+                raise ValueError("Must provide inputs to get output schema")
+
+            # TODO: Fix this properly
+            i = module.__input_schemas.index(inputs)
+
+            return output_schemas[i]
+
+        try:
+            return super().__call__(module, inputs)
+        except NotImplementedError:
+            raise ValueError(
+                f"Could not get output schema of {module} " "please call mm.trace_schema first."
+            )
+
+    def trace(
+        self,
+        module: nn.Module,
+        inputs: Union[torch.Tensor, Dict[str, torch.Tensor], Schema],
+        outputs: Union[torch.Tensor, Dict[str, torch.Tensor], Schema],
+    ) -> Schema:
+        _input_schema = input.get_schema(inputs)
+        _output_schema = self.get_schema(outputs)
+
+        try:
+            return super().__call__(module, _input_schema)
+        except NotImplementedError:
+            return _output_schema
+
+
+class _SelectDispatch(LazyDispatcher):
+    def __call__(self, to_select: ToSelectT, selection: Selection) -> ToSelectT:
+        if hasattr(to_select, "select") and not isinstance(to_select, Schema):
+            output = to_select.select(selection)
+        else:
+            output = super().__call__(to_select, selection)
+
+        return output
+
+
+class _ExtractDispatch(LazyDispatcher):
+    def __call__(self, module: nn.Module, selection: Selection) -> Tuple[nn.Module, nn.Module]:
+        extraction = select(module, selection)
+        module_with_extraction = self.extract(module, selection, extraction)
+
+        return module_with_extraction, extraction
+
+    def extract(self, module: nn.Module, selection: Selection, route: nn.Module, name=None):
+        fn = self.dispatch(module)
+        return fn(module, selection, route, name=name)
+
+
+input = _InputSchemaDispatch("input_schema")
+output = _OutputSchemaDispatch("output_schema")
+select = _SelectDispatch("selection")
+extract = _ExtractDispatch("extract")
+
+
+def trace(module: nn.Module, inputs: Union[torch.Tensor, Dict[str, torch.Tensor]], **kwargs):
+    """
+    Traces schemas for a given PyTorch module and all it's children.
+
+    Parameters
+    ----------
+    module : nn.Module
+        The PyTorch module to trace.
+    inputs : Union[torch.Tensor, Dict[str, torch.Tensor]]
+        The input to the module's forward method, it could be either a single Tensor
+        or a dictionary of Tensors.
+    **kwargs
+        Arbitrary keyword arguments to pass to the forward-pass.
+
+    Returns
+    -------
+    output : torch.Tensor
+        The output tensor of the module's forward pass.
+    """
+
+    hooks = []
+
+    def _hook(mod: nn.Module, inputs: Tuple[torch.Tensor], outputs: torch.Tensor):
+        if not hasattr(mod, "__input_schemas"):
+            mod.__input_schemas = ()
+            mod.__output_schemas = ()
+
+        _input_schema = input.trace(mod, inputs[0])
+        if _input_schema not in mod.__input_schemas:
+            mod.__input_schemas += (_input_schema,)
+            mod.__output_schemas += (output.trace(mod, _input_schema, outputs),)
+
+    def add_hook(m):
+        custom_modules = list(output.dispatcher.registry.keys())
+        if len(custom_modules) > 0 and isinstance(m, tuple(custom_modules[1:])):
+            return
+
+        hooks.append(m.register_forward_hook(_hook))
+
+    module.apply(add_hook)
+    module_out = module(inputs, **kwargs)
+
+    for hook in hooks:
+        hook.remove()
+
+    return module_out
+
+
+def setup_schema(module: nn.Module, schema: Schema):
+    """
+    Set up a schema for a given module.
+
+    Parameters
+    ----------
+    module : nn.Module
+        The module for which to set up the schema.
+    schema : Schema
+        The schema to set up.
+    """
+
+    from merlin.models.torch.block import BlockContainer, ParallelBlock
+
+    if hasattr(module, "setup_schema"):
+        module.setup_schema(schema)
+
+    elif isinstance(module, ParallelBlock):
+        for branch in module.branches.values():
+            setup_schema(branch, schema)
+
+    elif isinstance(module, BlockContainer) and module:
+        setup_schema(module[0], schema)
+
+
+@select.register(Schema)
+def select_schema(schema: Schema, selection: Selection) -> Schema:
+    """
+    Select a subset of a schema based on the selection criteria.
+
+    Parameters
+    ----------
+    schema : Schema
+        The original schema to select from.
+    selection : Selection
+        The selection criteria. Can be a Schema, ColumnSchema,
+        a callable that returns a Schema, or Tags.
+
+    Returns
+    -------
+    Schema
+        The selected subset of the schema.
+
+    Raises
+    ------
+    ValueError
+        If the selection criteria is not a valid type.
+    """
+    if not isinstance(schema, Schema):
+        raise ValueError(f"Schema {schema} is not valid")
+
+    if isinstance(selection, Schema):
+        selected = selection
+    elif isinstance(selection, ColumnSchema):
+        if selection.name not in schema.column_names:
+            return Schema()
+        selected = Schema([schema[selection.name]])
+    elif callable(selection):
+        selected = selection(schema)
+    elif isinstance(selection, Tags):
+        selected = schema.select_by_tag(selection)
+    else:
+        raise ValueError(f"Selection {selection} is not valid")
+
+    return selected
+
+
+def selection_name(selection: Selection) -> str:
+    """
+    Get the name of the selection.
+
+    Parameters
+    ----------
+    selection : Selection
+        The selection criteria. Can be a Schema, ColumnSchema, a callable
+        that returns a Schema, or Tags.
+
+    Returns
+    -------
+    str
+        The name of the selection.
+
+    Raises
+    ------
+    ValueError
+        If the selection criteria is not a valid type.
+    """
+    if isinstance(selection, ColumnSchema):
+        return selection.name
+    elif isinstance(selection, Tags):
+        return selection.value
+    elif isinstance(selection, Schema):
+        return "_".join(selection.column_names)
+    elif callable(selection):
+        return selection.__name__
+
+    raise ValueError(f"Selection {selection} is not valid")
+
+
+class Selectable:
+    """
+    A mixin to allow to be selectable by schema.
+    """
+
+    def setup_schema(self, schema: Schema):
+        """
+        Setup the schema for this selectable.
+
+        Parameters
+        ----------
+        schema : Schema
+            The schema to setup.
+
+        Returns
+        -------
+        Selectable
+            Self for chaining.
+        """
+        self.schema = schema
+
+        return self
+
+    def select(self, selection: Selection) -> "Selectable":
+        """
+        Select a subset of the schema.
+
+        This method should be overridden by subclasses.
+
+        Parameters
+        ----------
+        selection : Selection
+            The selection criteria. Can be a Schema, ColumnSchema, a callable that
+            returns a Schema, or Tags.
+
+        Returns
+        -------
+        Selectable
+            A new selectable with the selected subset of the schema.
+
+        Raises
+        ------
+        NotImplementedError
+            If this method is not overridden by a subclass.
+        """
+        raise NotImplementedError()
+
+
+@output.register_tensor(torch.Tensor)
+def _tensor_to_schema(input, name="output"):
+    kwargs = dict(dims=input.shape[1:], dtype=input.dtype)
+
+    if len(input.shape) > 1 and input.dtype != torch.int32:
+        kwargs["tags"] = [Tags.EMBEDDING]
+
+    return Schema([ColumnSchema(name, **kwargs)])
+
+
+@input.register_tensor(torch.Tensor)
+def _(input):
+    return _tensor_to_schema(input, "input")
+
+
+@input.register_tensor(Dict[str, torch.Tensor])
+@output.register_tensor(Dict[str, torch.Tensor])
+def _(input):
+    output = Schema()
+    for k, v in sorted(input.items()):
+        output += _tensor_to_schema(v, k)
+
+    return output
+
+
+@input.register_tensor(Tuple[torch.Tensor])
+@output.register_tensor(Tuple[torch.Tensor])
+@input.register_tensor(Tuple[torch.Tensor, torch.Tensor])
+@output.register_tensor(Tuple[torch.Tensor, torch.Tensor])
+@input.register_tensor(Tuple[torch.Tensor, torch.Tensor, torch.Tensor])
+@output.register_tensor(Tuple[torch.Tensor, torch.Tensor, torch.Tensor])
+@input.register_tensor(Tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor])
+@output.register_tensor(Tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor])
+def _(input):
+    output = Schema()
+
+    for i, v in enumerate(input):
+        output += _tensor_to_schema(v, str(i))
+
+    return output
